+++ apps/cli/src/live-common-setup-base.ts
@@ -98,6 +98,8 @@ setSupportedCurrencies([
   "casper",
   "neon_evm",
   "lukso",
+  "linea",
+  "linea_goerli",
 ]);
 
 for (const k in process.env) setEnvUnsafe(k as EnvName, process.env[k]);