+++ apps/ledger-live-desktop/src/renderer/components/Web3AppWebview/WalletAPIWebview.tsx
@@ -37,7 +37,7 @@ import { useWebviewState } from "./helpers";
 import { getStoreValue, setStoreValue } from "~/renderer/store";
 import { NetworkErrorScreen } from "./NetworkError";
 import getUser from "~/helpers/user";
-import { openExchangeDrawer, closePlatformAppDrawer } from "~/renderer/actions/UI";
+import { openExchangeDrawer } from "~/renderer/actions/UI";
 import { currentRouteNameRef } from "~/renderer/analytics/screenRefs";
 import { TrackFunction } from "@ledgerhq/live-common/platform/tracking";
 
@@ -171,11 +171,9 @@ function useUiHook(
             exchangeType: ExchangeType[exchangeType],
             onResult: (nonce: string) => {
               onSuccess(nonce);
-              dispatch(closePlatformAppDrawer());
             },
             onCancel: (error: Error) => {
               onCancel(error);
-              dispatch(closePlatformAppDrawer());
             },
           }),
         );
@@ -189,7 +187,6 @@ function useUiHook(
               onSuccess(operation.hash);
             },
             onCancel: (error: Error) => {
-              console.error(error);
               onCancel(error);
             },
           }),