+++ libs/ledger-live-common/src/exchange/platform/startExchange.ts
@@ -1,26 +1,28 @@
 import { log } from "@ledgerhq/logs";
-import { firstValueFrom, from } from "rxjs";
-import Exchange from "@ledgerhq/hw-app-exchange";
-import { Observable } from "rxjs";
+import { Observable, firstValueFrom, from } from "rxjs";
+import semver from "semver";
+import Exchange, { ExchangeTypes, isExchangeTypeNg } from "@ledgerhq/hw-app-exchange";
 import { withDevice } from "../../hw/deviceAccess";
 import type { ExchangeRequestEvent } from "../../hw/actions/startExchange";
+import { StartExchangeInput } from "./types";
+
 const withDevicePromise = (deviceId, fn) =>
   firstValueFrom(withDevice(deviceId)(transport => from(fn(transport))));
-
-type StartExchangeInput = {
-  deviceId: string;
-  exchangeType: number;
-};
 
 const startExchange = (input: StartExchangeInput): Observable<ExchangeRequestEvent> => {
   return new Observable(o => {
     let unsubscribed = false;
-    const { deviceId, exchangeType } = input;
+    const { deviceId, exchangeType, appVersion } = input;
 
     const startExchangeAsync = async () => {
       await withDevicePromise(deviceId, async transport => {
         log("exchange", `attempt to connect to ${deviceId}`);
         if (unsubscribed) return;
+
+        if (!checkNgPrerequisite(exchangeType, appVersion)) {
+          throw new Error("Incompatible AppExchange version with NG command");
+        }
+
         /**
          * Note: `transactionRate` is not needed at this stage. It is only used
          * at the `completeExchange` step
@@ -53,5 +55,19 @@ const startExchange = (input: StartExchangeInput): Observable<ExchangeRequestEve
     };
   });
 };
+
+// AppExchange Min version supporting NG types
+const MIN_APP_EXCHANGE_NG = "3.3.0";
+
+/**
+ * If the App-Exchnage version is not high enough, "NG" type can't worked
+ */
+function checkNgPrerequisite(exchangeType: ExchangeTypes, appVersion?: string) {
+  if (isExchangeTypeNg(exchangeType) && appVersion && semver.lt(appVersion, MIN_APP_EXCHANGE_NG)) {
+    return false;
+  }
+
+  return true;
+}
 
 export default startExchange;