+++ libs/ledgerjs/packages/cryptoassets/src/abandonseed.ts
@@ -108,6 +108,8 @@ const abandonSeedAddresses: Partial<Record<CryptoCurrency["id"], string>> = {
   dydx: "dydx19rl4cm2hmr8afy4kldpxz3fka4jguq0a4erelz",
   vechain: EVM_DEAD_ADDRESS,
   lukso: EVM_DEAD_ADDRESS,
+  linea: EVM_DEAD_ADDRESS,
+  linea_goerli: EVM_DEAD_ADDRESS,
 };
 
 /**