+++ libs/ledgerjs/packages/types-live/src/operation.ts
@@ -22,6 +22,9 @@ export type OperationType =
   | "FEES"
   | "FREEZE"
   | "UNFREEZE"
+  | "WITHDRAW_EXPIRE_UNFREEZE"
+  | "UNDELEGATE_RESOURCE"
+  | "LEGACY_UNFREEZE"
   // POLKADOT
   | "VOTE"
   | "REWARD_PAYOUT"