+++ libs/ui/examples/webpack.js/CHANGELOG.md
@@ -1,5 +1,19 @@
 # webpack.js-example
 
+## 0.0.10
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @ledgerhq/react-ui@0.14.13
+
+## 0.0.10-next.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @ledgerhq/react-ui@0.14.13-next.0
+
 ## 0.0.9
 
 ### Patch Changes